Check for 'psp' actually having a state
Turns out that up until now, the game would crash if the player became a zombie, specifically once the weapon finished lowering. This is because two of the conditions for calling P_NuggetBobbing() are checking for 'psp->state->action', but when the player's weapon gets lowered after dying, the state gets cleared. Simple solution: make sure there is a state before trying to reference data within it.

Additionally, changed P_NuggetBobbing() to simply calculate 'psp->sx2' and 'psp->sy2', since for the time being those are the only two values that must be calculated.