Check if 'crosshair.cr' is set before drawing it
It appears that, for some reason, when toggling the crosshair on for the first time with its toggle key HU_UpdateCrosshair() might be skipped altogether, therefore calling HU_DrawCrosshair() with a potentially uninitialized 'crosshair.cr', which causes a crash.

Seems to fix #19

As an alternative fix, a default 'crosshair.cr' could be set in HU_InitCrosshair() itself.

Additionally, when using the crosshair toggle key, only call HU_InitCrosshair() if the crosshair was actually enabled.