Change NUGHUD drawing position ranges
Both the max values for X and Y position have been reduced by 1, therefore making them 319 and 199 respectively, as 320 and 200 would draw the patch outside of the original view. Patches meant exclusively for widescreen can still be added through offsets in the patches, which would be necessary regardless of this change when it comes to the left side of the screen.